+++ rcgen/src/csr.rs
@@ -3,11 +3,14 @@ use std::hash::Hash;
 #[cfg(feature = "pem")]
 use pem::Pem;
 use pki_types::CertificateSigningRequestDer;
+use yasna::{models::ObjectIdentifier, DERWriter, Tag};
 
 #[cfg(feature = "pem")]
 use crate::ENCODE_CONFIG;
 use crate::{
-	Certificate, CertificateParams, Error, Issuer, PublicKeyData, SignatureAlgorithm, SigningKey,
+	key_pair::{serialize_public_key_der, sign_der},
+	oid, write_distinguished_name, write_x509_extension, Attribute, Certificate, CertificateParams,
+	Error, IsCa, Issuer, PublicKeyData, SignatureAlgorithm, SigningKey,
 };
 #[cfg(feature = "x509-parser")]
 use crate::{DistinguishedName, SanType};
@@ -43,6 +46,100 @@ pub struct CertificateSigningRequest {
 }
 
 impl CertificateSigningRequest {
+	/// Generate and serialize a certificate signing request (CSR) with custom PKCS #10 attributes.
+	/// as defined in [RFC 2986].
+	///
+	/// The constructed CSR will contain attributes based on the certificate parameters,
+	/// and include the subject public key information from `subject_key`. Additionally,
+	/// the CSR will be self-signed using the subject key.
+	///
+	/// Note that subsequent invocations of `serialize_request_with_attributes()` will not produce the exact
+	/// same output.
+	///
+	/// [RFC 2986]: <https://datatracker.ietf.org/doc/html/rfc2986#section-4>
+	pub fn new(
+		params: &CertificateParams,
+		subject_key: &impl SigningKey,
+		attrs: Vec<Attribute>,
+	) -> Result<Self, Error> {
+		// No .. pattern, we use this to ensure every field is used
+		#[deny(unused)]
+		let CertificateParams {
+			not_before,
+			not_after,
+			serial_number,
+			subject_alt_names,
+			distinguished_name,
+			is_ca,
+			key_usages,
+			extended_key_usages,
+			name_constraints,
+			crl_distribution_points,
+			custom_extensions,
+			use_authority_key_identifier_extension,
+			key_identifier_method,
+		} = params;
+		// - alg and key_pair will be used by the caller
+		// - not_before and not_after cannot be put in a CSR
+		// - key_identifier_method is here because self.write_extended_key_usage uses it
+		// - There might be a use case for specifying the key identifier
+		// in the CSR, but in the current API it can't be distinguished
+		// from the defaults so this is left for a later version if
+		// needed.
+		let _ = (
+			not_before,
+			not_after,
+			key_identifier_method,
+			extended_key_usages,
+		);
+		if serial_number.is_some()
+			|| *is_ca != IsCa::NoCa
+			|| name_constraints.is_some()
+			|| !crl_distribution_points.is_empty()
+			|| *use_authority_key_identifier_extension
+		{
+			return Err(Error::UnsupportedInCsr);
+		}
+
+		// Whether or not to write an extension request attribute
+		let write_extension_request = !key_usages.is_empty()
+			|| !subject_alt_names.is_empty()
+			|| !extended_key_usages.is_empty()
+			|| !custom_extensions.is_empty();
+
+		let der = sign_der(subject_key, |writer| {
+			// Write version
+			writer.next().write_u8(0);
+			write_distinguished_name(writer.next(), distinguished_name);
+			serialize_public_key_der(subject_key, writer.next());
+
+			// According to the spec in RFC 2986, even if attributes are empty we need the empty attribute tag
+			writer
+				.next()
+				.write_tagged_implicit(Tag::context(0), |writer| {
+					// RFC 2986 specifies that attributes are a SET OF Attribute
+					writer.write_set_of(|writer| {
+						if write_extension_request {
+							write_extension_request_attribute(params, writer.next());
+						}
+
+						for Attribute { oid, values } in attrs {
+							writer.next().write_sequence(|writer| {
+								writer.next().write_oid(&ObjectIdentifier::from_slice(&oid));
+								writer.next().write_der(&values);
+							});
+						}
+					});
+				});
+
+			Ok(())
+		})?;
+
+		Ok(Self {
+			der: CertificateSigningRequestDer::from(der),
+		})
+	}
+
 	/// Get the PEM-encoded bytes of the certificate signing request.
 	#[cfg(feature = "pem")]
 	pub fn pem(&self) -> Result<String, Error> {
@@ -63,6 +160,33 @@ impl From<CertificateSigningRequest> for CertificateSigningRequestDer<'static> {
 	fn from(csr: CertificateSigningRequest) -> Self {
 		csr.der
 	}
+}
+
+/// Write a CSR extension request attribute as defined in [RFC 2985].
+///
+/// [RFC 2985]: <https://datatracker.ietf.org/doc/html/rfc2985>
+pub(crate) fn write_extension_request_attribute(params: &CertificateParams, writer: DERWriter) {
+	writer.write_sequence(|writer| {
+		writer.next().write_oid(&ObjectIdentifier::from_slice(
+			oid::PKCS_9_AT_EXTENSION_REQUEST,
+		));
+		writer.next().write_set(|writer| {
+			writer.next().write_sequence(|writer| {
+				// Write key_usage
+				params.write_key_usage(writer.next());
+				// Write subject_alt_names
+				params.write_subject_alt_names(writer.next());
+				params.write_extended_key_usage(writer.next());
+
+				// Write custom extensions
+				for ext in &params.custom_extensions {
+					write_x509_extension(writer.next(), &ext.oid, ext.critical, |writer| {
+						writer.write_der(ext.content())
+					});
+				}
+			});
+		});
+	});
 }
 
 /// Parameters for a certificate signing request
